+++ src/test/java/com/iexec/core/replicate/ReplicateServiceTests.java
@@ -30,6 +30,8 @@ import com.iexec.core.result.ResultService;
 import io.vavr.control.Either;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.Mockito;
@@ -52,6 +54,8 @@ import java.util.stream.IntStream;
 
 import static com.iexec.common.replicate.ReplicateStatus.*;
 import static com.iexec.common.replicate.ReplicateStatusModifier.WORKER;
+import static com.iexec.commons.poco.chain.DealParams.DROPBOX_RESULT_STORAGE_PROVIDER;
+import static com.iexec.commons.poco.chain.DealParams.IPFS_RESULT_STORAGE_PROVIDER;
 import static com.iexec.commons.poco.utils.TestUtils.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
@@ -796,12 +800,14 @@ class ReplicateServiceTests {
 
     // region isResultUploaded
 
-    @Test
-    void shouldCheckResultServiceAndReturnTrue() {
+    @ParameterizedTest
+    @ValueSource(booleans = {true, false})
+    void shouldCheckResultServiceAndReturnTrue(boolean isTeeTask) {
         TaskDescription taskDescription = TaskDescription.builder()
                 .chainTaskId(CHAIN_TASK_ID)
                 .callback(BytesUtils.EMPTY_ADDRESS)
-                .isTeeTask(false)
+                .isTeeTask(isTeeTask)
+                .resultStorageProvider(IPFS_RESULT_STORAGE_PROVIDER)
                 .build();
         when(iexecHubService.getTaskDescription(CHAIN_TASK_ID))
                 .thenReturn(taskDescription);
@@ -812,12 +818,14 @@ class ReplicateServiceTests {
         verify(resultService).isResultUploaded(CHAIN_TASK_ID);
     }
 
-    @Test
-    void shouldCheckResultServiceAndReturnFalse() {
+    @ParameterizedTest
+    @ValueSource(booleans = {true, false})
+    void shouldCheckResultServiceAndReturnFalse(boolean isTeeTask) {
         TaskDescription taskDescription = TaskDescription.builder()
                 .chainTaskId(CHAIN_TASK_ID)
                 .callback(BytesUtils.EMPTY_ADDRESS)
-                .isTeeTask(false)
+                .isTeeTask(isTeeTask)
+                .resultStorageProvider(IPFS_RESULT_STORAGE_PROVIDER)
                 .build();
         when(iexecHubService.getTaskDescription(CHAIN_TASK_ID))
                 .thenReturn(taskDescription);
@@ -838,12 +846,13 @@ class ReplicateServiceTests {
         verify(resultService, never()).isResultUploaded(CHAIN_TASK_ID);
     }
 
-    @Test
-    void shouldReturnTrueForCallbackTask() {
+    @ParameterizedTest
+    @ValueSource(booleans = {true, false})
+    void shouldReturnTrueForCallbackTask(boolean isTeeTask) {
         TaskDescription taskDescription = TaskDescription.builder()
                 .chainTaskId(CHAIN_TASK_ID)
                 .callback("callback")
-                .isTeeTask(false)
+                .isTeeTask(isTeeTask)
                 .build();
         when(iexecHubService.getTaskDescription(CHAIN_TASK_ID))
                 .thenReturn(taskDescription);
@@ -854,18 +863,19 @@ class ReplicateServiceTests {
     }
 
     @Test
-    void shouldReturnTrueForTeeTask() {
+    void shouldReturnTrueIfPrivateStorageForTeeTask() {
         TaskDescription taskDescription = TaskDescription.builder()
                 .chainTaskId(CHAIN_TASK_ID)
                 .callback(BytesUtils.EMPTY_ADDRESS)
                 .isTeeTask(true)
+                .resultStorageProvider(DROPBOX_RESULT_STORAGE_PROVIDER)
                 .build();
         when(iexecHubService.getTaskDescription(CHAIN_TASK_ID))
                 .thenReturn(taskDescription);
 
         boolean isResultUploaded = replicatesService.isResultUploaded(CHAIN_TASK_ID);
         assertThat(isResultUploaded).isTrue();
-        verify(resultService, never()).isResultUploaded(CHAIN_TASK_ID);
+        verify(resultService, never()).isResultUploaded(any());
     }
     // endregion
 
@@ -992,7 +1002,7 @@ class ReplicateServiceTests {
                 .build();
         UpdateReplicateStatusArgs updateReplicateStatusArgs = UpdateReplicateStatusArgs
                 .builder()
-                .taskDescription(TaskDescription.builder().build())
+                .taskDescription(TaskDescription.builder().resultStorageProvider(IPFS_RESULT_STORAGE_PROVIDER).build())
                 .build();
 
         assertThat(replicatesService.canUpdateReplicateStatus(replicate, statusUpdate, updateReplicateStatusArgs))
@@ -1011,24 +1021,6 @@ class ReplicateServiceTests {
         UpdateReplicateStatusArgs updateReplicateStatusArgs = UpdateReplicateStatusArgs
                 .builder()
                 .taskDescription(TaskDescription.builder().callback("callback").build())
-                .build();
-
-        assertThat(replicatesService.canUpdateReplicateStatus(replicate, statusUpdate, updateReplicateStatusArgs))
-                .isEqualTo(ReplicateStatusUpdateError.GENERIC_CANT_UPDATE);
-    }
-
-    @Test
-    void shouldNotAuthorizeUpdateOnResultUploadFailedSinceResultUploadedWithTee() {
-        Replicate replicate = new Replicate(WALLET_WORKER_1, CHAIN_TASK_ID);
-        replicate.updateStatus(RESULT_UPLOADING, ReplicateStatusModifier.WORKER);
-
-        ReplicateStatusUpdate statusUpdate = ReplicateStatusUpdate.builder()
-                .modifier(WORKER)
-                .status(RESULT_UPLOAD_FAILED)
-                .build();
-        UpdateReplicateStatusArgs updateReplicateStatusArgs = UpdateReplicateStatusArgs
-                .builder()
-                .taskDescription(TaskDescription.builder().isTeeTask(true).build())
                 .build();
 
         assertThat(replicatesService.canUpdateReplicateStatus(replicate, statusUpdate, updateReplicateStatusArgs))
@@ -1229,6 +1221,7 @@ class ReplicateServiceTests {
                 .thenReturn(true);
         when(iexecHubService.getTaskDescription(CHAIN_TASK_ID)).thenReturn(task);
         when(iexecHubService.isTaskInCompletedStatusOnChain(CHAIN_TASK_ID)).thenReturn(true);
+        when(resultService.isResultUploaded(CHAIN_TASK_ID)).thenReturn(true);
 
         assertThat(replicatesService.canUpdateReplicateStatus(replicate, statusUpdate, null))
                 .isEqualTo(ReplicateStatusUpdateError.NO_ERROR);
@@ -1261,7 +1254,11 @@ class ReplicateServiceTests {
                 .modifier(WORKER)
                 .status(CONTRIBUTE_AND_FINALIZE_DONE)
                 .build();
-        final TaskDescription task = TaskDescription.builder().chainTaskId(CHAIN_TASK_ID).build();
+        final TaskDescription task = TaskDescription
+                .builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .resultStorageProvider(IPFS_RESULT_STORAGE_PROVIDER)
+                .build();
 
         when(iexecHubService.repeatIsRevealedTrue(CHAIN_TASK_ID, WALLET_WORKER_1))
                 .thenReturn(true);